Connect safeArea margin value with redux
